malloc protection, small corrections

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: allefebv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/06 11:20:09 by allefebv          #+#    #+#             */
-/*   Updated: 2018/11/09 14:36:34 by allefebv         ###   ########.fr       */
+/*   Updated: 2018/11/13 18:38:31 by allefebv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	*ft_memmove(void *dst, const void *src, size_t n)
 {
 	unsigned char	*dst_2;
 	unsigned char	*src_2;
-	unsigned long	i;
+	size_t			i;
 
 	dst_2 = (unsigned char*)dst;
 	src_2 = (unsigned char*)src;
